cr: lots of new comments, rename helper module to aml for clarity
