+++ vm/acpi/src/aml/helpers.rs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+//! Utilities for encoding various types into ACPI Machine Language (AML).
+
 pub fn encode_name(name: &[u8]) -> Vec<u8> {
     let mut encoded_name: Vec<u8> = Vec::new();
     let mut segments: Vec<[u8; 4]> = Vec::new();
@@ -153,7 +155,7 @@ pub fn char_to_hex(value: u8) -> u8 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::sdt::test_helpers::verify_expected_bytes;
+    use crate::aml::test_helpers::verify_expected_bytes;
 
     #[test]
     fn verify_simple_name() {
