+++ vm/acpi/src/lib.rs
@@ -6,7 +6,7 @@
 #![expect(missing_docs)]
 #![forbid(unsafe_code)]
 
+mod aml;
 pub mod builder;
 pub mod dsdt;
-mod sdt;
 pub mod ssdt;
